+++ packages/sdk/src/evm/contracts/prebuilt-implementations/nft-drop.ts
@@ -44,7 +44,7 @@ import { BigNumber, BigNumberish, CallOverrides, constants } from "ethers";
  * import { ThirdwebSDK } from "@thirdweb-dev/sdk";
  *
  * const sdk = new ThirdwebSDK("{{chainName}}");
- * const contract = sdk.getContract("{{contract_address}}", "nft-drop");
+ * const contract = await sdk.getContract("{{contract_address}}", "nft-drop");
  * ```
  *
  * @public