+++ packages/sdk/src/evm/contracts/prebuilt-implementations/split.ts
@@ -30,7 +30,7 @@ import { BigNumber, CallOverrides, Contract } from "ethers";
  * import { ThirdwebSDK } from "@thirdweb-dev/sdk";
  *
  * const sdk = new ThirdwebSDK("{{chainName}}");
- * const contract = sdk.getContract("{{contract_address}}", "split");
+ * const contract = await sdk.getContract("{{contract_address}}", "split");
  * ```
  *
  * @public